+++ gwcelery/tests/test_nagios.py
@@ -1,5 +1,5 @@
 from distutils.spawn import find_executable
-from unittest.mock import patch
+from unittest.mock import Mock
 
 import pytest
 
@@ -7,10 +7,12 @@ from .. import app
 from .. import nagios
 
 
-@patch('gwcelery.app.connection', side_effect=RuntimeError)
-def test_nagios_unknown_error(mock_connection, capsys):
+def test_nagios_unknown_error(monkeypatch, capsys):
     """Test that we generate the correct message when there is an unexpected
     exception."""
+    monkeypatch.setattr('gwcelery.app.connection',
+                        Mock(side_effect=RuntimeError))
+
     with pytest.raises(SystemExit) as excinfo:
         app.start(['gwcelery', 'nagios'])
     assert excinfo.value.code == nagios.NagiosPluginStatus.UNKNOWN
@@ -18,7 +20,11 @@ def test_nagios_unknown_error(mock_connection, capsys):
     assert 'UNKNOWN: Unexpected error' in out
 
 
-def test_nagios(capsys, monkeypatch, socket_enabled, starter, tmpdir):
+def test_nagios(capsys, monkeypatch, socket_enabled,
+                starter, tmpdir):
+    mock_lvalert_client = Mock()
+    monkeypatch.setattr('sleek_lvalert.LVAlertClient', mock_lvalert_client)
+
     # no broker
 
     unix_socket = str(tmpdir / 'sock')
@@ -58,3 +64,40 @@ def test_nagios(capsys, monkeypatch, socket_enabled, starter, tmpdir):
     assert excinfo.value.code == nagios.NagiosPluginStatus.CRITICAL
     out, err = capsys.readouterr()
     assert 'CRITICAL: Not all expected tasks are active' in out
+
+    # tasks, no LVAlert nodes
+
+    monkeypatch.setattr('gwcelery.nagios.get_active_tasks',
+                        lambda _: nagios.get_expected_tasks(app))
+    mock_lvalert_client.configure_mock(**{
+        'return_value.get_subscriptions.return_value': {}})
+
+    with pytest.raises(SystemExit) as excinfo:
+        app.start(['gwcelery', 'nagios'])
+    assert excinfo.value.code == nagios.NagiosPluginStatus.CRITICAL
+    out, err = capsys.readouterr()
+    assert 'CRITICAL: Not all lvalert nodes are subscribed' in out
+
+    # tasks, too many LVAlert nodes
+
+    mock_lvalert_client.configure_mock(**{
+        'return_value.get_subscriptions.return_value':
+        nagios.get_expected_lvalert_nodes() | {'foobar'}})
+
+    with pytest.raises(SystemExit) as excinfo:
+        app.start(['gwcelery', 'nagios'])
+    assert excinfo.value.code == nagios.NagiosPluginStatus.CRITICAL
+    out, err = capsys.readouterr()
+    assert 'CRITICAL: Too many lvalert nodes are subscribed' in out
+
+    # success
+
+    mock_lvalert_client.configure_mock(**{
+        'return_value.get_subscriptions.return_value':
+        nagios.get_expected_lvalert_nodes()})
+
+    with pytest.raises(SystemExit) as excinfo:
+        app.start(['gwcelery', 'nagios'])
+    assert excinfo.value.code == nagios.NagiosPluginStatus.OK
+    out, err = capsys.readouterr()
+    assert 'OK: Running normally' in out