+++ CHANGES.md
@@ -10,6 +10,9 @@
 -   Allow detchar module's `create_cache` to use gwdatafind when frames
     are no longer in llhoft.
 
+-   The Nagios monitoring plugin will now report on the status of LVAlert
+    subscriptions.
+
 ## 0.2.4 (2018-12-17)
 
 -   Fix broken links in log messages due to changes in GraceDb URL routes.